DDL: createTable

+++ final/src/main/java/com/example/client/Class.java
@@ -12,7 +12,7 @@ import com.example.annotation.Table;
 @Table(name = "classes")
 public class Class {
     @Id
-    @Column(name = "id", type = JDBCType.BIGINT)
+    @Column(name = "id", type = JDBCType.INTEGER)
     private int id;
 
     @Column(name = "name")
@@ -24,7 +24,7 @@ public class Class {
     @OneToMany(mappedBy = "classObject") // Back reference for Users
     private List<User> users = new ArrayList<>(); // Initialize to avoid null issues
 
-    // Getters and Setters
+    //Getters and Setters
     public int getId() {
         return id;
     }
